put almost everything into folders

+++ src/gui/cycle_button.hpp
@@ -97,7 +97,7 @@ void CycleButton<T>::handleMouseClick(float x, float y) {
 
 template <typename T>
 size_t CycleButton<T>::getIndex(T t) {
-	for (int i = 0; i < _entries.size(); ++i) {
+	for (size_t i = 0; i < _entries.size(); ++i) {
 		if (_entries[i].val == t)
 			return i;
 	}
